Feat/network configuration

+++ server-agent/server_agent.go
@@ -1,12 +1,18 @@
 package main
 
 import (
+	"bufio"
 	"embed"
+	"errors"
 	"log"
+	"net/http"
 	"os"
 	"os/exec"
+	"path/filepath"
+	"runtime"
 	"strconv"
 	"strings"
+	"syscall"
 	"text/template"
 	"time"
 )
@@ -30,13 +36,14 @@ type ServerAgent interface {
 	StopInstance(instanceId string) error
 	RestartInstance(instanceId string) error
 	ListInstancesStatus() ([]ListInstancesStatusResponse, error)
+	GetResourceStatus() (GetResourceStatusResponse, error)
 }
 
 type ServerAgentImpl struct {
-	vmsStoragePath       string
-	cloudInitImagesPath  string
-	defaultNetworkBridge string
-	vmNetworkBridge      string
+	vmsStoragePath      string
+	cloudInitImagesPath string
+	vmsBridge           string
+	vmNetworkInterface  string
 }
 
 type VmType string
@@ -84,8 +91,6 @@ type CloudInitNetworkConfig struct {
 }
 
 func (agent *ServerAgentImpl) ListBaseImages() (ListBaseImagesResponse, error) {
-	log.Printf("Getting base images...")
-
 	cmd := exec.Command(
 		"ls",
 		agent.cloudInitImagesPath,
@@ -155,6 +160,17 @@ func (agent *ServerAgentImpl) DeleteVm(request DeleteVmRequest) error {
 	output, err := cmd.CombinedOutput()
 
 	if err != nil {
+		// If the domain doesn't exist, it means it does not exist in this server
+		// we need to remove the vlan etiquete from the network bridge anyway
+		if strings.Contains(string(output), "failed to get domain") {
+			if request.RemoveEtiquete {
+				if err := agent.removeVidFromNetworkBridge(request.Vid); err != nil {
+					return err
+				}
+			}
+			return nil
+		}
+
 		return logAndReturnError("Error deleting VM '"+request.VmId+"': ", string(output))
 	}
 
@@ -163,8 +179,6 @@ func (agent *ServerAgentImpl) DeleteVm(request DeleteVmRequest) error {
 	if err := os.RemoveAll(agent.vmsStoragePath + "/" + request.VmId); err != nil {
 		return logAndReturnError("Error deleting VM '"+request.VmId+"' files from storage: ", err.Error())
 	}
-
-	log.Printf("VM '%s' files removed from storage successfully!", request.VmId)
 
 	log.Printf("Deleted VM '%s' successfully!", request.VmId)
 
@@ -179,6 +193,12 @@ func (agent *ServerAgentImpl) DeleteVm(request DeleteVmRequest) error {
 
 func (agent *ServerAgentImpl) StartInstance(request StartInstanceRequest) error {
 	log.Printf("Starting instance '%s'...", request.InstanceId)
+
+	if !agent.vmDomainExists(request.InstanceId) {
+		if err := agent.importVmDomain(request.InstanceId); err != nil {
+			return err
+		}
+	}
 
 	cmd := exec.Command(
 		"virsh", "start", request.InstanceId,
@@ -206,6 +226,12 @@ func (agent *ServerAgentImpl) StopInstance(instanceId string) error {
 	)
 
 	if output, err := cmd.CombinedOutput(); err != nil {
+		// If the domain doesn't exist, or is already shut down, we return a bad request
+		// to inform the vms-manager that the instance is not running in this server
+		if strings.Contains(string(output), "failed to get domain") || strings.Contains(string(output), "Domain is not running") {
+			return NewHttpError(http.StatusBadRequest, errors.New("instance is not running in this server"))
+		}
+
 		return logAndReturnError("Error stopping instance '"+instanceId+"': ", string(output))
 	}
 
@@ -224,7 +250,6 @@ func (agent *ServerAgentImpl) StopInstance(instanceId string) error {
 		for _, status := range statuses {
 			if status.InstanceId == instanceId {
 				if status.Status != SHUTOFF_STATUS {
-					log.Printf("Time since current time: %s, retryShutdown: %t", time.Since(currentTime), retryShutdown)
 					if time.Since(currentTime) >= FORCE_SHUTDOWN_WAIT_TIME {
 						return agent.forceStopVM(instanceId)
 					} else if time.Since(currentTime) >= RETRY_SHUTDOWN_WAIT_TIME && retryShutdown {
@@ -250,6 +275,12 @@ func (agent *ServerAgentImpl) StopInstance(instanceId string) error {
 func (agent *ServerAgentImpl) RestartInstance(instanceId string) error {
 	log.Printf("Restarting instance '%s'...", instanceId)
 
+	if !agent.vmDomainExists(instanceId) {
+		// If the domain doesn't exist, we return a bad request to inform the vms-manager
+		// that the instance is not running in this server
+		return NewHttpError(http.StatusBadRequest, errors.New("instance is not running in this server"))
+	}
+
 	cmd := exec.Command(
 		"virsh", "reboot", instanceId,
 	)
@@ -257,6 +288,12 @@ func (agent *ServerAgentImpl) RestartInstance(instanceId string) error {
 	output, err := cmd.CombinedOutput()
 
 	if err != nil {
+		// If the domain is not running, we return a bad request to inform the vms-manager
+		// that the instance is not running in this server
+		if strings.Contains(string(output), "Domain is not running") {
+			return NewHttpError(http.StatusBadRequest, errors.New("instance is not running in this server"))
+		}
+
 		return logAndReturnError("Error restarting instance '"+instanceId+"': ", string(output))
 	}
 
@@ -266,8 +303,6 @@ func (agent *ServerAgentImpl) RestartInstance(instanceId string) error {
 }
 
 func (agent *ServerAgentImpl) ListInstancesStatus() ([]ListInstancesStatusResponse, error) {
-	log.Printf("Getting VMs status...")
-
 	cmd := exec.Command(
 		"virsh", "list", "--all",
 	)
@@ -297,9 +332,32 @@ func (agent *ServerAgentImpl) ListInstancesStatus() ([]ListInstancesStatusRespon
 	return toListInstancesStatusResponse(vmStatusMap), nil
 }
 
-func (agent *ServerAgentImpl) createVm(request CreateVmRequest) error {
-	log.Printf("Creating %s '%s'...", request.VmType, request.VmId)
+func (agent *ServerAgentImpl) GetResourceStatus() (GetResourceStatusResponse, error) {
+	cpuLoad, err := getCpuLoad()
+	if err != nil {
+		return GetResourceStatusResponse{}, err
+	}
 
+	totalMemoryMB, freeMemoryMB, err := getMemoryInfo()
+	if err != nil {
+		return GetResourceStatusResponse{}, err
+	}
+
+	totalDiskMB, freeDiskMB, err := getDiskInfo()
+	if err != nil {
+		return GetResourceStatusResponse{}, err
+	}
+
+	return GetResourceStatusResponse{
+		CpuLoad:       cpuLoad,
+		TotalMemoryMB: totalMemoryMB,
+		FreeMemoryMB:  freeMemoryMB,
+		TotalDiskMB:   totalDiskMB,
+		FreeDiskMB:    freeDiskMB,
+	}, nil
+}
+
+func (agent *ServerAgentImpl) createVm(request CreateVmRequest) error {
 	if err := createDir(request.DirPath); err != nil {
 		return err
 	}
@@ -326,7 +384,8 @@ func (agent *ServerAgentImpl) createVm(request CreateVmRequest) error {
 		return err
 	}
 
-	log.Printf("%s '%s' created successfully!", request.VmType, request.VmId)
+	agent.dumpVmXML(request.VmId)
+
 	return nil
 }
 
@@ -334,14 +393,12 @@ func (agent *ServerAgentImpl) removeVidFromNetworkBridge(vid string) error {
 	log.Printf("Removing vlan etiquete from network bridge...")
 
 	removeVlanEtiqueteCmd := exec.Command(
-		"bridge", "vlan", "del", "vid", vid, "dev", agent.vmNetworkBridge, "self", // TODO: Remove self on production
+		"bridge", "vlan", "del", "vid", vid, "dev", agent.vmNetworkInterface,
 	)
 
 	if output, err := removeVlanEtiqueteCmd.CombinedOutput(); err != nil {
 		return logAndReturnError("Error removing vlan etiquete from network bridge: ", string(output))
 	}
-
-	log.Printf("Vlan etiquete removed from network bridge successfully!")
 
 	return nil
 }
@@ -350,7 +407,7 @@ func (agent *ServerAgentImpl) setupVMNetwork(vid string, vlanEtiquete string) er
 	log.Printf("Setting up network...")
 
 	addVlanEtiqueteCmd := exec.Command(
-		"bridge", "vlan", "add", "vid", vid, "dev", agent.vmNetworkBridge, "self", // TODO: Remove self on production
+		"bridge", "vlan", "add", "vid", vid, "dev", agent.vmNetworkInterface,
 	)
 
 	if output, err := addVlanEtiqueteCmd.CombinedOutput(); err != nil {
@@ -364,8 +421,6 @@ func (agent *ServerAgentImpl) setupVMNetwork(vid string, vlanEtiquete string) er
 	if output, err := setupVmInterfaceAsAccessPortCmd.CombinedOutput(); err != nil {
 		return logAndReturnError("Error setting up vm interface as access port: ", string(output))
 	}
-
-	log.Printf("Network setup successfully!")
 
 	return nil
 }
@@ -382,7 +437,6 @@ func createDir(dirPath string) error {
 		return logAndReturnError("Error creating directory: ", err.Error())
 	}
 
-	log.Printf("Directory '%s' created successfully!", dirPath)
 	return nil
 }
 
@@ -443,7 +497,6 @@ func (agent *ServerAgentImpl) createVmConfigurationFiles(request CreateVmRequest
 		return err
 	}
 
-	log.Printf("%s configuration files created successfully!", request.VmType)
 	return nil
 }
 
@@ -466,7 +519,6 @@ func createFileFromTemplate(newFilePath string, fileName string, data interface{
 
 	file.Close()
 
-	log.Printf("File %s created successfully!", fileName)
 	return nil
 }
 
@@ -493,7 +545,6 @@ func createCidataIso(dirPath string, isTemplate bool) error {
 		return logAndReturnError("Error creating cidata.iso: ", string(createIsoCmdOutput))
 	}
 
-	log.Printf("Cidata.iso created successfully!")
 	return nil
 }
 
@@ -522,7 +573,6 @@ func (agent *ServerAgentImpl) createDiskImage(request CreateVmRequest) error {
 		return logAndReturnError("Error creating disk image: ", string(createDiskImageCmdOutput))
 	}
 
-	log.Printf("Disk image created successfully!")
 	return nil
 }
 
@@ -544,7 +594,6 @@ func (agent *ServerAgentImpl) removeBackingFileFromTemplateDiskImage(dirPath str
 		)
 	}
 
-	log.Printf("Backing file removed from template disk image successfully!")
 	return nil
 }
 
@@ -560,7 +609,7 @@ func (agent *ServerAgentImpl) installVm(request CreateVmRequest) error {
 		"--disk", "path="+request.DirPath+"/"+request.VmId+".qcow2,format=qcow2",
 		"--disk", "path="+request.DirPath+"/cidata.iso,device=cdrom",
 		"--os-variant", DEFAULT_OS_VARIANT,
-		"--network", "bridge="+agent.defaultNetworkBridge+",target="+request.VlanEtiquete+",model=virtio",
+		"--network", "bridge="+agent.vmsBridge+",target="+request.VlanEtiquete+",model=virtio",
 		"--graphics", "vnc,listen=0.0.0.0",
 		"--noautoconsole",
 	)
@@ -570,12 +619,34 @@ func (agent *ServerAgentImpl) installVm(request CreateVmRequest) error {
 		return logAndReturnError("Error installing VM: ", string(installVmCmdOutput))
 	}
 
-	log.Printf("VM installed successfully!")
-
 	// We need to wait for the VM to be started and configured
 	time.Sleep(AFTER_INSTALL_WAIT_TIME)
 
 	return nil
+}
+
+func (agent *ServerAgentImpl) dumpVmXML(vmId string) {
+	xmlPath := filepath.Join(agent.vmsStoragePath, vmId, vmId+".xml")
+
+	log.Printf("Dumping VM XML to %s...", xmlPath)
+
+	for {
+		cmd := exec.Command("virsh", "dumpxml", vmId)
+		output, err := cmd.Output()
+		if err != nil {
+			log.Printf("Error dumping VM XML: %v; retrying…", err)
+			time.Sleep(time.Second)
+			continue
+		}
+
+		if err := os.WriteFile(xmlPath, output, 0644); err != nil {
+			log.Printf("Error writing XML file: %v; retrying…", err)
+			time.Sleep(time.Second)
+			continue
+		}
+
+		return
+	}
 }
 
 func (agent *ServerAgentImpl) forceStopVM(vmId string) error {
@@ -591,7 +662,34 @@ func (agent *ServerAgentImpl) forceStopVM(vmId string) error {
 		return logAndReturnError("Error force stopping VM '"+vmId+"': ", string(output))
 	}
 
-	log.Printf("Force stopped VM '%s' successfully!", vmId)
+	return nil
+}
+
+func (agent *ServerAgentImpl) vmDomainExists(vmId string) bool {
+	domains, err := agent.ListInstancesStatus()
+	if err != nil {
+		return false
+	}
+
+	for _, domain := range domains {
+		if domain.InstanceId == vmId {
+			return true
+		}
+	}
+
+	return false
+}
+
+func (agent *ServerAgentImpl) importVmDomain(vmId string) error {
+	log.Printf("Importing VM domain...")
+
+	importVmDomainCmd := exec.Command(
+		"virsh", "define", agent.vmsStoragePath+"/"+vmId+"/"+vmId+".xml",
+	)
+
+	if output, err := importVmDomainCmd.CombinedOutput(); err != nil {
+		return logAndReturnError("Error importing VM domain: ", string(output))
+	}
 
 	return nil
 }
@@ -604,16 +702,95 @@ func toListInstancesStatusResponse(vmStatusMap map[string]string) []ListInstance
 	return response
 }
 
+func getCpuLoad() (float64, error) {
+	data, err := os.ReadFile("/proc/loadavg")
+	if err != nil {
+		return 0, err
+	}
+
+	fields := strings.Fields(string(data))
+	loadAvgString := fields[0] // 1-minute load average
+
+	loadAvg, err := strconv.ParseFloat(loadAvgString, 64)
+	if err != nil {
+		return 0, err
+	}
+
+	cpuCount := runtime.NumCPU()
+
+	cpuLoad := loadAvg / float64(cpuCount)
+
+	return cpuLoad, nil
+}
+
+func getMemoryInfo() (totalMemoryMB int, freeMemoryMB int, err error) {
+	totalMemoryMB = -1
+	freeMemoryMB = -1
+
+	file, err := os.Open("/proc/meminfo")
+	if err != nil {
+		return 0, 0, err
+	}
+	defer file.Close()
+
+	scanner := bufio.NewScanner(file)
+	for scanner.Scan() {
+		line := scanner.Text()
+		fields := strings.Fields(line)
+
+		if len(fields) < 2 {
+			continue
+		}
+
+		key := strings.TrimSuffix(fields[0], ":")
+		value, err2 := strconv.Atoi(fields[1])
+		if err2 != nil {
+			continue
+		}
+
+		if key == "MemFree" {
+			// Return in MB
+			freeMemoryMB = value / 1024
+		} else if key == "MemTotal" {
+			totalMemoryMB = value / 1024
+		}
+	}
+
+	if totalMemoryMB == -1 || freeMemoryMB == -1 {
+		return 0, 0, logAndReturnError("MemFree or MemTotal not found", "")
+	}
+
+	return totalMemoryMB, freeMemoryMB, nil
+}
+
+func getDiskInfo() (totalDiskMB int, freeDiskMB int, err error) {
+	var stat syscall.Statfs_t
+
+	if err = syscall.Statfs("/", &stat); err != nil {
+		return
+	}
+
+	// Blocks * size-per-block = bytes
+	total := stat.Blocks * uint64(stat.Bsize)
+	free := stat.Bavail * uint64(stat.Bsize)
+
+	// Convert to MB
+	totalDiskMB = int(total / (1024 * 1024))
+	freeDiskMB = int(free / (1024 * 1024))
+
+	return totalDiskMB, freeDiskMB, nil
+}
+
 func NewServerAgent(
 	vmsStoragePath string,
 	cloudInitImagesPath string,
-	defaultNetworkBridge string,
-	vmNetworkBridge string,
+	vmsBridge string,
+	vmNetworkInterface string,
 ) ServerAgent {
 	return &ServerAgentImpl{
-		vmsStoragePath:       vmsStoragePath,
-		cloudInitImagesPath:  cloudInitImagesPath,
-		defaultNetworkBridge: defaultNetworkBridge,
-		vmNetworkBridge:      vmNetworkBridge,
+		vmsStoragePath:      vmsStoragePath,
+		cloudInitImagesPath: cloudInitImagesPath,
+		vmsBridge:           vmsBridge,
+		vmNetworkInterface:  vmNetworkInterface,
 	}
 }
